keep values in nbx when you hit enter
If you click on [nbx], type a value and hit enter, the value is output, but if you hit enter again, it empties the value.

Also, when you click on it and hit enter without any new entry, the existing value becomes zero as well.

partially closes https://github.com/pure-data/pure-data/issues/1747