more improvements to 16.partialtracer.pd
This closes https://github.com/pure-data/pddp/issues/115

and adds mor improvements. I added a bit more information on how you can manually edit the analysis result and refer to a subpatch that was hidden and allows you to generate traces. I am now also using a subpatch that was hidden that synthesizes a sample.