Update sigmund~.c
this removes the currently unused loud argument to the list input because it is not really working and is mandatory. I'm doing this because of the discussion in https://github.com/pure-data/pure-data/issues/1847 - more specifically in https://github.com/pure-data/pure-data/issues/1847#issuecomment-1337672944

If this is accepted I will update the documentation accordingly to remove the flag as a mandatory argument.