Have the active outlet-area spread over the entire object-width
similar to how inlets are handled.
so if you start dragging from "somewhere" (x-wise) at the bottom of the object,
this will start a connection from the the nearest outlet,
even if you were not directly above the outlet.

this also fixes the start-position of the the connection-line to match the
start position of the finalized connection.

Related: https://github.com/pure-data/pure-data/issues/1926