document "print" message in [sigmund~]
we missed that one and seems like a good thing to have and document.

there's also a 'printnext' message that I don't understand and need the help of @millerpuckette

other minor cosmetic changes to the other files here. In partialtracer and FFT plot I have introduced a hack so the main window is visible and 'display' window stays in the background. I also used '$0' in the sample name.