i18n: handle the case where the system requests a set of multiple languages
typically a set of multiple languages (e.g. "fr_CA:en_CA") defines a preferred
language with fallbacks (use "French (Canada)", and if that's not available
use "English (Canada)".

however, the current handling is naive and just use the first
language (here "fr_CA"). this is mostly because the actual loading code
in ::pd_i18n::load_locale does not know which languages are available-
LATER unify this