taskqueue: keep a list of pending tasks
this simplifies the logic in task_stop() and taskqueue_stop() a lot.

Also, task_stop() is now safe because we can check whether the given task is alive!
(For example, an external might accidentally call task_stop() on a stale task handle.)
task_stop() now returns 1 on success and 0 if the task could not be found.