readsf~/writesf~: add multichannel support
readsf~: if created with the "-m" flag, the object only has a single multichannel outlet;
the channel count can be changed dynamically with the "channels" method.

writesf~: the object uses all channels of all input signals.
We post a warning if the channel count changes while the object is streaming;
this can also happen when the "open" message is sent before DSP is turned on.