improve/fix sigmund~ help
the 'print' message was updated for 0.08 but not the 'post' when loading the object, so I also changed this in the code.

The default values for quality and amppower were wrong, I fixed it.

We had too much repeating information, so I reorganized things a bit and left the details for parameters in the already existing subpatch for setting the parameters.

I added a new example subptach with 'notes' output, which also uses 'quality' and added information from the release notes about the purpose of 'evenness'.