make [savestate] work in cloned abstractions
in clone_save() we ask all abstractions to save their state ("#A saved ...") to a temporary binbuf.
if the state is not empty, we prepend it with a "#A set <i>" message which will set the target abstraction instance.
when the clone object is reloaded, it will receive all the "saved" messages and forward them to the corresponding (= currently set) abstraction.