clone: allow to create anonymous abstractions
If you supply "pd" as the abstraction name, clone will create an anonymous abstraction.
The only notable difference to regular abstraction is that it is not saved as a seperate file, instead it is embedded in the clone object's owning canvas.
A special hack also allows to retext the clone object, e.g. changing the number of instances or creation arguments, without losing the abstraction data.
This is particularly useful with the new multichannel functionality, in particular the "-d" flag.