Make sure that DEKEN_OS and DEKEN_CPU are passed as quoted defines
otherwise we see weird effects if the *value* is also defined as a macro.
e.g. compiling with `-DDEKEN_CPU="i386"` will pass the literal `i386`
to the preprocessor-engine. On x86_32 (aka "i386"), gcc will have a predefined
macro 'i386' that is set to 1.
`stringify(DEKEN_CPU)` will then generate a string "1" and we end up with
a bogus deken specifier like "Linux-1-64".

so we have to pass it double-quoted, e.g. `-DDEKEN_CPU="'i386'"`, which will
then expand to the string "'i386'" (with the extra quotes), which we then
have to strip down again.

Closes: https://github.com/pure-data/pure-data/issues/2101