nbx: add a 'disable clicks' message
This tries to follow the semantics of Max's number box, using the same message name cantchange, and identical (from my testing) semantics.

This is not fully Max-compatible yet, as it applies only to nbx/my_numbox, not floatatoms, and does not add support for the parameter in the properties TCL properties page, or initialization arguments/attributes.