deal with different channels size
I am using a logic described in here https://github.com/pure-data/pure-data/issues/2107#issuecomment-1774156090

If you have a multichannel connection in the left inlet and secondary inlets are just a single channel, then they are copied and used to clip the left input. Now, if the secondary inlets have more than a single channel, then it needs to be the same size of the mutichannel left input, otherwise we have an error and zeroes are output.