Remove EXTERN specifier from function definitions (as opposed to declarations)
the public interface (including EXTERN) should (only) live in the header-files.

the only exception is error(), which we no longer expose in the headers
but still export.

this also declares mayer_dofft() (in src/d_fft_fftsg.c) as "static",
as it really is an implementation detail. the public API is mayer_fft() resp mayer_ifft()