document default flag values in [soundfiler] and letting it save files
closes https://github.com/pure-data/pddp/issues/177

One big issue is that the example used a '/tmp/' directory to write to, but this directory doesn't exist, so noting got saved. I removed this temp file so now it gets saved on the docs folder.

I also documented default flag values for writting

I think it would best to also use [openpanel] and [savepanel] in the examples by the way.