fix 'pd_this' on Windows
On Windows it is not possible to 'properly' export a thread local variable from a DLL.
MinGW can do it, but the resulting symbol ('__emutls_v.pd_this') is not portable and certainly not understood by MSVC.

The solution is to access 'pd_this' via a function call.
For externals, 'pd_this' is now a macro for 'pd_getinstance()'.
Internally, we directly use 'pd_this', but we don't try to export it.

On other platforms, 'pd_this' is now properly marked as EXTERN.