readsf: add third outlet for soundfile info
when the user opens a soundfile with "open" and the file has been opened successfully,
we send the soundfile info through the third outlet.

actually, we also wait until the buffer has been filled sufficiently, so the message
can be used to determine when the object is ready for playback.
In particular, the user does not have to guess the wait time between "open" and "start" anymore.