first attempt at reimplementing text fields in scalars, and also to allow "editing" inside atom boxes insige GOPs.
To make inside-GOP-editing work it was necessary to move all "rtexts" in GOPs
out to the owning canvas ala glist_getcanvas(..).  This required making
lots of changes, often changing "glist" to glist_getcanvas(glist).  There
are probably spots where this is still not done correctly.

clicking away from atom boxes (and text fields in scalars) now adopts
the new text instead of resetting to the previous value (as it was
before.)  Instead, revert-to-previous is done by hitting "undo".  But
this isn't done using the full-on undo stack, it's merely canceled with
no 'redo' possible.  It's not clear what actions should be undo-able.  A
first idea would be "everything done in edit mode" but there should be
undo-availability when editing data structures.

But... what when a performer draws a shap into an array at run-time, for
example to shance a waveform?  It sounds wrong to slap that onto the
undo stack.  Some clear thinking is needed :)

There is a minor rendering bug that pops up when you double click on a
number/symbol/list box in run mode.  I'm leaving that for later since
I'm still thinking about editing scalars and will probably need to make
more changes to g_editor.c .