improve description of [line~] restrictions
text now says "Note that the \[line~\] is always bound to audio block boundaries. This means it always starts a ramp in the beginning of an audio block and finishes it at the end of an audio block as well. If you want better precision with ramps that are shorter than an audio block or ramps that start and stop in the middle of an audio blocks you can use \[vline~\], which also has a delay parameter to schedule different ramps."

closes https://github.com/pure-data/pddp/issues/270