feat: Add support to `oscparse` for handling `d` type tag (double)
The Open Sound Control 1.0 spec mentions a (non-standard) `d` type tag
for handling 64-bit floating point values (`double`). This adds support
to the `oscparse` object for parsing osc args of this type.

Note that this implementation results in a potentially lossy conversion
from `double` to `t_float` in the case that `t_float` is `float`.

This was motivated by some personal software that emits doubles for
certain control values. Previously I was receiving these in
supercollider without issue, but after experimenting with receiving the
same data in puredata I noticed the `d` tag was unsupported, so thought
I'd have a go at contributing it!