have iemguis accept "default" as a color, for future compatibility
at some point we'd like to be able to pick "default" colors based on Pd's
current color "scheme".

with this patch, we fill in some better values than the default "black".

note: we use the same default color for all iemguis, which means that
canvas and vu get the wrong background color.