make the "clear" canvas message safe
We defer destruction of all non-scalar objects with canvas_defer().
This prevents a crash when (accidentally) clearing the containing canvas (or a parent canvas) during message passing.

Scalars are still deleted synchronously, otherwise we would break existing data structure patches.
This is actually safe because scalars do not participate in message passing and are protected by gpointers.